Button component Refactoring / simplify your user data fetching code with recoil
